+++ pkg/analysis_server/lib/src/channel/byte_stream_channel.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-// @dart = 2.9
-
 import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
@@ -17,7 +15,6 @@ import 'package:analyzer/instrumentation/instrumentation.dart';
 /// [ClientCommunicationChannel] that uses a stream and a sink (typically,
 /// standard input and standard output) to communicate with servers.
 class ByteStreamClientChannel implements ClientCommunicationChannel {
-  final Stream input;
   final IOSink output;
 
   @override
@@ -26,22 +23,36 @@ class ByteStreamClientChannel implements ClientCommunicationChannel {
   @override
   Stream<Notification> notificationStream;
 
-  ByteStreamClientChannel(this.input, this.output) {
-    Stream jsonStream = input
+  factory ByteStreamClientChannel(Stream<List<int>> input, IOSink output) {
+    var jsonStream = input
         .transform(const Utf8Decoder())
         .transform(LineSplitter())
         .transform(JsonStreamDecoder())
-        .where((json) => json is Map)
+        .where((json) => json is Map<String, Object?>)
+        .cast<Map<String, Object?>>()
         .asBroadcastStream();
-    responseStream = jsonStream
+    var responseStream = jsonStream
         .where((json) => json[Notification.EVENT] == null)
         .transform(ResponseConverter())
+        .where((response) => response != null)
+        .cast<Response>()
         .asBroadcastStream();
-    notificationStream = jsonStream
+    var notificationStream = jsonStream
         .where((json) => json[Notification.EVENT] != null)
         .transform(NotificationConverter())
         .asBroadcastStream();
+    return ByteStreamClientChannel._(
+      output,
+      responseStream,
+      notificationStream,
+    );
   }
+
+  ByteStreamClientChannel._(
+    this.output,
+    this.responseStream,
+    this.notificationStream,
+  );
 
   @override
   Future close() {
@@ -69,7 +80,7 @@ class ByteStreamServerChannel implements ServerCommunicationChannel {
   final InstrumentationService _instrumentationService;
 
   /// The helper for recording request / response statistics.
-  final RequestStatisticsHelper _requestStatistics;
+  final RequestStatisticsHelper? _requestStatistics;
 
   /// Completer that will be signalled when the input stream is closed.
   final Completer _closed = Completer();
@@ -79,7 +90,7 @@ class ByteStreamServerChannel implements ServerCommunicationChannel {
 
   ByteStreamServerChannel(
       this._input, this._output, this._instrumentationService,
-      {RequestStatisticsHelper requestStatistics})
+      {RequestStatisticsHelper? requestStatistics})
       : _requestStatistics = requestStatistics {
     _requestStatistics?.serverChannel = this;
   }
@@ -100,12 +111,12 @@ class ByteStreamServerChannel implements ServerCommunicationChannel {
 
   @override
   void listen(void Function(Request request) onRequest,
-      {Function onError, void Function() onDone}) {
+      {Function? onError, void Function()? onDone}) {
     _input.transform(const Utf8Decoder()).transform(LineSplitter()).listen(
         (String data) => _readRequest(data, onRequest),
         onError: onError, onDone: () {
       close();
-      onDone();
+      onDone?.call();
     });
   }
 
@@ -148,7 +159,7 @@ class ByteStreamServerChannel implements ServerCommunicationChannel {
 
   /// Read a request from the given [data] and use the given function to handle
   /// the request.
-  void _readRequest(Object data, void Function(Request request) onRequest) {
+  void _readRequest(String data, void Function(Request request) onRequest) {
     // Ignore any further requests after the communication channel is closed.
     if (_closed.isCompleted) {
       return;