+++ pkg/analysis_server/test/services/search/hierarchy_test.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-// @dart = 2.9
-
 import 'package:analysis_server/src/services/search/hierarchy.dart';
 import 'package:analysis_server/src/services/search/search_engine_internal.dart';
 import 'package:analyzer/dart/element/element.dart';
@@ -20,7 +18,7 @@ void main() {
 
 @reflectiveTest
 class HierarchyTest extends AbstractSingleUnitTest {
-  SearchEngineImpl searchEngine;
+  late SearchEngineImpl searchEngine;
 
   @override
   void setUp() {
@@ -73,7 +71,7 @@ class B extends A {
     var futureB = getHierarchyMembers(searchEngine, memberB).then((members) {
       expect(members, unorderedEquals([memberB]));
     });
-    return Future.wait([futureA, futureB]);
+    await Future.wait([futureA, futureB]);
   }
 
   Future<void> test_getHierarchyMembers_fields() async {
@@ -111,7 +109,7 @@ class D {
     var futureD = getHierarchyMembers(searchEngine, memberD).then((members) {
       expect(members, unorderedEquals([memberD]));
     });
-    return Future.wait([futureA, futureB, futureC, futureD]);
+    await Future.wait([futureA, futureB, futureC, futureD]);
   }
 
   Future<void> test_getHierarchyMembers_fields_static() async {
@@ -189,7 +187,7 @@ class E extends D {
     var futureE = getHierarchyMembers(searchEngine, memberE).then((members) {
       expect(members, unorderedEquals([memberD, memberE]));
     });
-    return Future.wait([futureA, futureB, futureC, futureD, futureE]);
+    await Future.wait([futureA, futureB, futureC, futureD, futureE]);
   }
 
   Future<void> test_getHierarchyMembers_methods_static() async {
@@ -247,7 +245,7 @@ class E {
     var futureD = getHierarchyMembers(searchEngine, memberD).then((members) {
       expect(members, unorderedEquals([memberA, memberB, memberD]));
     });
-    return Future.wait([futureA, futureB, futureD]);
+    await Future.wait([futureA, futureB, futureD]);
   }
 
   Future<void> test_getHierarchyNamedParameters() async {
@@ -403,7 +401,7 @@ class F implements A {}
     var classD = findElement.class_('D');
     var classE = findElement.class_('E');
     var classF = findElement.class_('F');
-    var objectElement = classA.supertype.element;
+    var objectElement = classA.supertype!.element;
     // Object
     {
       var supers = getSuperClasses(objectElement);
@@ -461,7 +459,7 @@ mixin M5 on A, C {}
     var m3 = findElement.mixin('M3');
     var m4 = findElement.mixin('M4');
     var m5 = findElement.mixin('M5');
-    var object = a.supertype.element;
+    var object = a.supertype!.element;
 
     _assertSuperClasses(object, []);
     _assertSuperClasses(a, [object]);