+++ pkg/analysis_server/test/src/services/correction/fix/data_driven/transform_set_parser_test_support.dart
@@ -42,7 +42,8 @@ abstract class AbstractTransformSetParserTest {
 
   void parse(String content) {
     errorListener = GatheringErrorListener();
-    var errorReporter = ErrorReporter(errorListener, MockSource('data.yaml'));
+    var errorReporter =
+        ErrorReporter(errorListener, MockSource(fullName: 'data.yaml'));
     var parser = TransformSetParser(errorReporter, 'myPackage');
     result = parser.parse(content);
   }